services/agent/requirements.txt
# ============================================================================
# Nyx Venatrix Agent Dependencies
# Python: 3.12 | Embeddings: OpenAI-only | Agents: Ray
# ============================================================================

# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20

# Browser Automation
browser-use==0.9.7
cdp-use==1.4.4
playwright==1.56.0
beautifulsoup4==4.14.2
lxml==6.0.2
markdownify==1.2.2

# Multi-Agent Concurrency
ray[default]==2.40.0

# LLM & Embeddings (OpenAI-only)
openai==2.8.1
tiktoken==0.12.0
langchain==1.0.8
langchain-core==1.0.7
langchain-openai==1.0.3
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.45

# Alternative LLM Providers
anthropic==0.74.1
groq==0.36.0
google-genai==1.52.0
ollama==0.6.1

# Observability & Tracking
mlflow==2.19.0
langfuse>=2.0.0
prometheus-client==0.21.0

# Database
psycopg2-binary==2.9.11
SQLAlchemy[postgresql]==2.0.44
qdrant-client==1.16.0

# Vector Store & Embeddings
# Note: Using OpenAI embeddings exclusively, no sentence-transformers

# Utilities
aiofiles==25.1.0
requests==2.32.5
httpx==0.28.1
aiohttp==3.12.15
backoff==2.2.1
tenacity==9.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Document Processing
pypdf==6.3.0
python-docx==1.2.0
reportlab==4.4.5

# Telegram Integration
python-telegram-bot>=20.0

# CAPTCHA Solving
# Note: 2captcha integration via HTTP API

# Data Processing
pandas==2.3.3
numpy==2.3.5

# Scientific Computing (minimal, no PyTorch)
scikit-learn==1.7.2
scipy==1.16.3

# Authentication & Security
cryptography==46.0.3
pyotp==2.9.0
Authlib==1.6.5

# CLI & UI
inquirerpy==0.3.4
click==8.3.1
rich==14.2.0
prompt_toolkit==3.0.52
pfzy==0.3.4

# Google APIs
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0

# MCP (Model Context Protocol)
mcp==1.12.4

# Development & Testing
pytest==8.0.0
pytest-asyncio==0.23.0

# Misc Dependencies
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bubus==1.5.6
cachetools>=5.0.0
certifi==2025.11.12
cffi>=1.15.0
charset-normalizer==3.4.4
cloudpickle==3.1.2
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
oauthlib==3.3.1
orjson==3.11.4
ormsgpack==1.12.0
packaging>=23.0
pillow==12.0.0
portalocker==2.10.1
posthog==7.0.1
propcache==0.4.1
proto-plus==1.26.1
protobuf>=4.25.0
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.29.0
rsa==4.9.1
screeninfo==0.8.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.3
starlette==0.35.1
sympy==1.14.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
